changed uint to unsigned, changed uint8_t to unsigned

+++ log.h
@@ -118,13 +118,13 @@ typedef struct _log_sink {
    * \param data sink data
    */
   void (*dispose)(void *data);
-  uint8_t filter;
+  int filter;
 } log_sink;
 
 typedef struct _log_sink_list {
   log_sink *sinks;
-  uint8_t   sink_count;
-  uint8_t   main_filter;
+  unsigned  sink_count;
+  unsigned  main_filter;
 } log_sink_list;
 
 
@@ -243,19 +243,19 @@ int  log_to_syslog_priority(enum LogSeverity severity);
 /**\brief close all sinks by calling \link log_sink#dispose dispose callback
  * \endlink
  */
-#define LOGGER_SHUTDOWN()                                                  \
-  {                                                                        \
-    for (uint log_sink_counter = 0; log_sink_counter < LOGGER->sink_count; \
-         ++log_sink_counter) {                                             \
-      log_sink *sink = &LOGGER->sinks[log_sink_counter];                   \
-      if (sink->dispose) {                                                 \
-        sink->dispose(sink->data);                                         \
-      }                                                                    \
-    }                                                                      \
-    free(LOGGER->sinks);                                                   \
-    LOGGER->sinks       = NULL;                                            \
-    LOGGER->sink_count  = 0;                                               \
-    LOGGER->main_filter = 0;                                               \
+#define LOGGER_SHUTDOWN()                                                      \
+  {                                                                            \
+    for (unsigned log_sink_counter = 0; log_sink_counter < LOGGER->sink_count; \
+         ++log_sink_counter) {                                                 \
+      log_sink *sink = &LOGGER->sinks[log_sink_counter];                       \
+      if (sink->dispose) {                                                     \
+        sink->dispose(sink->data);                                             \
+      }                                                                        \
+    }                                                                          \
+    free(LOGGER->sinks);                                                       \
+    LOGGER->sinks       = NULL;                                                \
+    LOGGER->sink_count  = 0;                                                   \
+    LOGGER->main_filter = 0;                                                   \
   }
 
 #define LOG_PREPARE_MESSAGE(severity, ...)                          \
@@ -275,7 +275,7 @@ int  log_to_syslog_priority(enum LogSeverity severity);
                           const char *     function,                         \
                           const char *     log_fmt_str_buf,                  \
                           void *           data) = NULL;                                \
-  for (uint log_sink_counter = 0; log_sink_counter < LOGGER->sink_count;     \
+  for (unsigned log_sink_counter = 0; log_sink_counter < LOGGER->sink_count; \
        ++log_sink_counter) {                                                 \
     log_sink *cur_log_sink = &LOGGER->sinks[log_sink_counter];               \
     if (cur_log_sink->filter & (severity)) {                                 \
